Changed a Setter's Args to Double to match Models Member Field Types

+++ src/main/java/com/truckstopservices/inventory/fuel/model/FuelModel.java
@@ -57,7 +57,7 @@ public abstract class FuelModel {
         return availableGallons;
     }
 
-    public void setAvailableGallons(int availableGallons) {
+    public void setAvailableGallons(double availableGallons) {
         this.availableGallons = availableGallons;
     }
